# apps/server/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions": {
      "post": {
        "operationId": "SessionController_createSession",
        "parameters": [],
        "responses": { "201": { "description": "Session created successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Create a new session",
        "tags": ["sessions"]
      }
    },
    "/api/sessions/{id}/upload": {
      "post": {
        "operationId": "SessionController_uploadAudio",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "type": "object", "properties": { "audio": { "type": "string", "format": "binary" } } }
            }
          }
        },
        "responses": { "201": { "description": "Audio uploaded successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Upload audio for a session",
        "tags": ["sessions"]
      }
    },
    "/api/sessions/{id}/status": {
      "put": {
        "operationId": "SessionController_updateSessionStatus",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Session status updated successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Update session status",
        "tags": ["sessions"]
      }
    },
    "/api/sessions/practitioner/{id}": {
      "get": {
        "operationId": "SessionController_getSessionsByPractitioner",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Sessions retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get sessions by practitioner",
        "tags": ["sessions"]
      }
    },
    "/api/sessions/practitioner/me": {
      "get": {
        "operationId": "SessionController_getMySessions",
        "parameters": [],
        "responses": { "200": { "description": "Sessions retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get sessions for current practitioner",
        "tags": ["sessions"]
      }
    },
    "/api/sessions/client/{id}": {
      "get": {
        "operationId": "SessionController_getSessionsByClient",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Sessions retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get sessions by client",
        "tags": ["sessions"]
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "operationId": "SessionController_getSessionById",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Session retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get session by ID",
        "tags": ["sessions"]
      }
    },
    "/api/plans": {
      "post": {
        "operationId": "PlanController_createPlan",
        "parameters": [],
        "responses": { "201": { "description": "Plan created successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Create a new plan",
        "tags": ["plans"]
      }
    },
    "/api/plans/{id}/publish": {
      "patch": {
        "operationId": "PlanController_publishPlan",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Plan published successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Publish a plan",
        "tags": ["plans"]
      }
    },
    "/api/plans/practitioner/{id}": {
      "get": {
        "operationId": "PlanController_getPlansByPractitioner",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Plans retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get plans by practitioner",
        "tags": ["plans"]
      }
    },
    "/api/plans/client/{id}": {
      "get": {
        "operationId": "PlanController_getPlansByClient",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Plans retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get plans by client",
        "tags": ["plans"]
      }
    },
    "/api/plans/{id}": {
      "get": {
        "operationId": "PlanController_getPlanById",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Plan retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get plan by ID",
        "tags": ["plans"]
      },
      "put": {
        "operationId": "PlanController_updatePlan",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Plan updated successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Update a plan",
        "tags": ["plans"]
      }
    },
    "/api/plans/{id}/with-suggestions": {
      "get": {
        "operationId": "PlanController_getPlanWithSuggestions",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Plan with suggestions retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get plan with AI suggestions",
        "tags": ["plans"]
      }
    },
    "/api/plans/{id}/suggestions": {
      "get": {
        "operationId": "PlanController_getSuggestedActionItems",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Suggested action items retrieved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Get suggested action items for a plan",
        "tags": ["plans"]
      }
    },
    "/api/plans/suggestions/{suggestionId}/approve": {
      "post": {
        "operationId": "PlanController_approveSuggestedActionItem",
        "parameters": [{ "name": "suggestionId", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Suggested action item approved successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Approve a suggested action item",
        "tags": ["plans"]
      }
    },
    "/api/plans/suggestions/{suggestionId}/reject": {
      "post": {
        "operationId": "PlanController_rejectSuggestedActionItem",
        "parameters": [{ "name": "suggestionId", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Suggested action item rejected successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Reject a suggested action item",
        "tags": ["plans"]
      }
    },
    "/api/plans/suggestions/{suggestionId}": {
      "patch": {
        "operationId": "PlanController_updateSuggestedActionItem",
        "parameters": [{ "name": "suggestionId", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Suggested action item updated successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Edit a suggested action item",
        "tags": ["plans"]
      }
    },
    "/api/plans/{id}/action-items": {
      "post": {
        "operationId": "PlanController_addCustomActionItem",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Custom action item added successfully." } },
        "security": [{ "bearer": [] }],
        "summary": "Add a custom action item to a plan",
        "tags": ["plans"]
      }
    },
...
  },
  "info": { "title": "NestJS API", "description": "API for NestJS", "version": "1.0", "contact": {} },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
...
  }
}